requirements.txt
flask>=3.0.0
httpx>=0.27.0
PyJWT[crypto]>=2.8.0
cryptography>=42.0.0
python-dotenv>=1.0.0
APScheduler>=3.10.0
mcp[cli]>=1.2.0
google-auth>=2.30.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.130.0
google-cloud-storage>=2.17.0
